requirements.txt
kubernetes
pprint
argparse
slackclient